+++ src/battle_anim_water.c
@@ -893,7 +893,7 @@ void AnimTask_CreateSurfWave(u8 taskId)
 static void AnimTask_CreateSurfWave_Step1(u8 taskId)
 {
     struct BattleAnimBgData animBg;
-    u8 i;
+    u32 i;
     u16 rgbBuffer;
     u16 *BGptrX = &gBattle_BG1_X;
     u16 *BGptrY = &gBattle_BG1_Y;
@@ -1151,7 +1151,7 @@ static void AnimTask_WaterSpoutLaunch_Step(u8 taskId)
 // A higher number results in more water sprites during the Water Spout animation
 static u8 GetWaterSpoutPowerForAnim(void)
 {
-    u8 i;
+    u32 i;
     u16 hp;
     u16 maxhp;
     u16 partyIndex;
@@ -1478,7 +1478,7 @@ static void AnimWaterSportDroplet(struct Sprite *sprite)
 
 static void AnimWaterSportDroplet_Step(struct Sprite *sprite)
 {
-    u16 i;
+    u32 i;
 
     if (TranslateAnimHorizontalArc(sprite))
     {
