+++ src/battle_interface.c
@@ -1183,7 +1183,7 @@ void UpdateHpTextInHealthbox(u8 healthboxSpriteId, s16 value, u8 maxOrCurrent)
         else
         {
             u32 var;
-            u8 i;
+            u32 i;
 
             if (GetBattlerSide(gSprites[healthboxSpriteId].data[6]) == B_SIDE_PLAYER)
             {
@@ -1261,7 +1261,7 @@ static void UpdateHpTextInHealthboxInDoubles(u8 healthboxSpriteId, s16 value, u8
             u8 var = 4;
             u8 r7;
             u8 *txtPtr;
-            u8 i;
+            u32 i;
 
             if (maxOrCurrent == HP_CURRENT)
                 var = 0;
@@ -1314,7 +1314,7 @@ static void PrintSafariMonInfo(u8 healthboxSpriteId, struct Pokemon *mon)
     u8 text[20];
     s32 j, spriteTileNum;
     u8 *barFontGfx;
-    u8 i, var, nature, healthBarSpriteId;
+    u32 i, var, nature, healthBarSpriteId;
 
     memcpy(text, sEmptyWhiteText_GrayHighlight, sizeof(sEmptyWhiteText_GrayHighlight));
     barFontGfx = &gMonSpritesGfxPtr->barFontGfx[0x520 + (GetBattlerPosition(gSprites[healthboxSpriteId].hMain_Battler) * 384)];
@@ -2277,7 +2277,7 @@ static void MoveBattleBarGraphically(u8 battler, u8 whichBar)
     u8 array[8];
     u8 filledPixelsCount, level;
     u8 barElementId;
-    u8 i;
+    u32 i;
 
     switch (whichBar)
     {
@@ -2413,7 +2413,7 @@ static s32 CalcNewBarValue(s32 maxValue, s32 oldValue, s32 receivedValue, s32 *c
 static u8 CalcBarFilledPixels(s32 maxValue, s32 oldValue, s32 receivedValue, s32 *currValue, u8 *pixelsArray, u8 scale)
 {
     u8 pixels, filledPixels, totalPixels;
-    u8 i;
+    u32 i;
 
     s32 newValue = oldValue - receivedValue;
     if (newValue < 0)
@@ -2483,7 +2483,7 @@ static void Debug_TestHealthBar_Helper(struct TestingBar *barInfo, s32 *currValu
 {
     u8 pixels[6];
     u16 src[6];
-    u8 i;
+    u32 i;
 
     CalcBarFilledPixels(barInfo->maxValue, barInfo->oldValue,
                 barInfo->receivedValue, currValue, pixels, B_HEALTHBAR_PIXELS / 8);
