+++ src/battle_tower.c
@@ -1823,7 +1823,7 @@ static void FillFactoryTrainerParty(void)
 
 static void FillFactoryFrontierTrainerParty(u16 trainerId, u8 firstMonId)
 {
-    u8 i, j;
+    u32 i, j;
     u8 friendship;
     u8 level;
     u8 fixedIV;
@@ -1886,7 +1886,7 @@ static void FillFactoryFrontierTrainerParty(u16 trainerId, u8 firstMonId)
 
 static void FillFactoryTentTrainerParty(u16 trainerId, u8 firstMonId)
 {
-    u8 i, j;
+    u32 i, j;
     u8 friendship;
     u8 level = TENT_MIN_LEVEL;
     u8 fixedIV = 0;
@@ -3470,7 +3470,7 @@ static void FillTentTrainerParty_(u16 trainerId, u8 firstMonId, u8 monCount)
 u8 FacilityClassToGraphicsId(u8 facilityClass)
 {
     u8 trainerObjectGfxId;
-    u8 i;
+    u32 i;
 
     // Search male classes.
     for (i = 0; i < ARRAY_COUNT(gTowerMaleFacilityClasses); i++)
