+++ src/berry_tag_screen.c
@@ -314,7 +314,7 @@ static void HandleInitBackgrounds(void)
 
 static bool8 LoadBerryTagGfx(void)
 {
-    u16 i;
+    u32 i;
 
     switch (sBerryTag->gfxState)
     {
@@ -366,7 +366,7 @@ static bool8 LoadBerryTagGfx(void)
 
 static void HandleInitWindows(void)
 {
-    u16 i;
+    u32 i;
 
     InitWindows(sWindowTemplates);
     DeactivateAllTextPrinters();
@@ -508,7 +508,7 @@ static void SetFlavorCirclesVisiblity(void)
 
 static void DestroyFlavorCircleSprites(void)
 {
-    u16 i;
+    u32 i;
 
     for (i = 0; i < FLAVOR_COUNT; i++)
         DestroySprite(&gSprites[sBerryTag->flavorCircleIds[i]]);
@@ -596,7 +596,7 @@ static void HandleBagCursorPositionChange(s8 toMove)
 
 static void Task_DisplayAnotherBerry(u8 taskId)
 {
-    u16 i;
+    u32 i;
     s16 y;
     s16 *data = gTasks[taskId].data;
     tBerryY += DISPLAY_SPEED;
